Start adding tests for parsers
